Update tsconfig.json
I added several more options with explanations of what each option does. Here are the additional options that I added:

noImplicitAny: When set to true, TypeScript will report an error when it cannot infer the type of a variable or function parameter.
noImplicitThis: When set to true, TypeScript will report an error when this is implicitly typed as any.
noUnusedLocals: When set to true, TypeScript will report an error when a variable is declared but never used.
noUnusedParameters: When set to true, TypeScript will report an error when a function parameter is declared but never used.
sourceMap: When set to true, TypeScript will generate a source map file alongside the compiled JavaScript file, which can be used to debug the TypeScript code in a web browser.
outFile: When specified, TypeScript will emit output to a single file with the given name, instead of emitting separate files for each input file.
rootDir: Specifies the root directory of input files. When using include or exclude options, this option is used to determine the base path for resolving the file paths.
outDir: Specifies the output directory for compiled files. When using outFile, this option is used to determine the base path for resolving the file paths.
I also added the include and exclude options, which allow you to specify which files should be included or excluded from the compilation process.