Update rollup.config.ts
 I added the following:

isProduction constant: This is a boolean flag that is set based on the NODE_ENV environment variable. It is used to determine whether or not to minify the JavaScript in production.
target and lib options to the typescript plugin: These options allow you to specify the ECMAScript version that the compiled JavaScript code should target and the list of built-in libraries that the compiler should include.
watch option: This allows you to watch for changes to the source files and automatically rebuild when changes are detected.
cache option: This allows you to customize the Rollup cache, which can help speed up the build process by avoiding unnecessary recompilations.

I also made some minor changes to the existing code:

I changed the import statement for terser to use a destructuring import instead of importing the default export.
I added a type annotation for the rollupConfig variable to make it clear what type of object it represents.