Update jest.config.ts
I added the following:

testRunner option: This specifies the test runner to use. Jest supports several runners, including the default jest-jasmine and the newer jest-circus. jest-circus is recommended for larger projects because it provides better test isolation and performance.
testMatch option: This specifies the file extensions to include in the tests. In this example, we're including all files with the .test.ts and .test.tsx extensions in the src directory.
moduleNameMapper option: This allows you to map module names to paths, which can be useful for importing mocks and stubs. In this example, we're mapping the @ and ~ prefixes to the src and test directories, respectively.
testPathIgnorePatterns option: This allows you to ignore files and directories when searching for tests. In this example, we're ignoring the node_modules directory.
collectCoverage option: This enables Jest's code coverage feature, which measures how much of your code is being tested. The coverage information is output to the coverage directory.
testEnvironment option: This specifies the environment in which the tests will be run. In this example, we're using the Node.js environment.
verbose and maxWorkers options: These options allow you to customize the output of the test results. In this example, we're setting verbose to true to get more detailed output, and maxWorkers to 4 to parallelize the tests across multiple CPU cores.
watchPathIgnorePatterns option: This allows you to ignore certain files and directories when watching for changes. In this example, we're ignoring the node_modules and .git directories.
clearMocks option: This clears the mock calls between test runs, which can help prevent test leakage.